Next release (`v3.1.0`)

+++ R/mod_reweightView.R
@@ -26,7 +26,7 @@ reweightViewUI <- function(id) {
 #' @keywords internal
 reweightViewServer <- function(id, user, decision_list, credentials) {
   if (missing(credentials))
-    credentials <- get_db_config("credentials")
+    credentials <- get_credential_config()
   moduleServer(id, function(input, output, session) {
     
     exportTestValues(
@@ -46,7 +46,7 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
       )
     
     observeEvent(input$update_weight, {
-      req("weight_adjust" %in% credentials$privileges[[user$role]])
+      req("weight_adjust" %in% unlist(credentials$privileges[user$role], use.names = FALSE))
       curr_new_wts(save$data %>%
                      dplyr::mutate(new_weight = ifelse(name == isolate(input$metric_name),
                                                        isolate(input$metric_weight), new_weight)))
@@ -93,11 +93,11 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
               div(class = "box-body",
                   br(),
                   fluidRow(
-                    column(width = 5, offset = 5, align = "left",
+                    column(width = 5, offset = 4, align = "left",
                            h3("Set new weights:"),
                     )),
                   fluidRow(
-                    column(width = 2, offset = 5, align = "left",
+                    column(width = 2, offset = 4, align = "left",
                            selectInput(NS(id, "metric_name"), "Select metric", curr_new_wts()$name, selected = curr_new_wts()$name[1]) ),
                     column(width = 2, align = "left",
                            numericInput(NS(id, "metric_weight"), "Choose new weight", min = 0, value = curr_new_wts()$new_weight[1]) ),
@@ -106,7 +106,7 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
                            actionButton(NS(id, "update_weight"), "Confirm", class = "btn-secondary") ) ),
                   br(), br(), 
                   fluidRow(
-                    column(width = 3, offset = 1, align = "center",
+                    column(width = 3, offset = 0, align = "center",
                            
                            br(), br(), br(), 
                            tags$hr(class = "hr_sep"),
@@ -126,7 +126,7 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
                            
                     ),
                     column(width = 6, style = "border: 1px solid rgb(77, 141, 201)",
-                           offset = 1,
+                           offset = 0,
                            h3("Current Risk Score Weights by Metric", align = "center"),
                            DT::dataTableOutput(NS(id, "weights_table")))
                   ),
@@ -184,7 +184,7 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
     
     # Update metric weight dropdown so that it matches the metric name.
     observeEvent(input$metric_name, {
-      req("weight_adjust" %in% credentials$privileges[[user$role]])
+      req("weight_adjust" %in% unlist(credentials$privileges[user$role], use.names = FALSE))
       
       shinyjs::disable("update_weight")
       updateNumericInput(session, "metric_weight",
@@ -199,14 +199,14 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
     # Note that another of the observeEvents will update the metric weight after
     # the selected metric name is updated.
     observeEvent(input$weights_table_rows_selected, {
-      req("weight_adjust" %in% credentials$privileges[[user$role]])
+      req("weight_adjust" %in% unlist(credentials$privileges[user$role], use.names = FALSE))
       updateSelectInput(session, "metric_name",
                         selected = curr_new_wts()$name[input$weights_table_rows_selected])
     })
     
     # Save new weight into db.
     observeEvent(input$update_pkg_risk, {
-      req("weight_adjust" %in% credentials$privileges[[user$role]])
+      req("weight_adjust" %in% unlist(credentials$privileges[user$role], use.names = FALSE))
       
       # if you the user goes input$back2dash, then when they return to the 
       if(n_wts_chngd() == 0){
@@ -225,7 +225,10 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
                tags$ul(
                  tags$li("The new package weights will be applied and risk metric scores re-calculated."),
                  tags$li("The risk re-calculation will be logged as a comment for each package."),
-                 tags$li("Previously finalized decisions & overall comments will be dropped for re-evaluation.")
+                 checkboxInput(NS(id, "reset_decisions"), width = "100%", value = isolate(input$reset_decisions) %||% FALSE,
+                               "Previously finalized decisions & overall comments will be dropped for re-evaluation.") %>%
+                   tagAppendAttributes(style = "position: absolute; left: 1rem; margin: 0",
+                                       .cssSelector = "input")
                )
             ),
             h3(strong("Note:"), "Updating the risk metrics cannot be reverted.", class = "mt-25 mb-0"),
@@ -243,7 +246,7 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
     
     # Upon confirming the risk re-calculation
     observeEvent(input$confirm_update_risk, {
-      req("weight_adjust" %in% credentials$privileges[[user$role]])
+      req("weight_adjust" %in% unlist(credentials$privileges[user$role], use.names = FALSE))
       removeModal()
       
       session$userData$trigger_events[["reset_pkg_upload"]] <- session$userData$trigger_events[["reset_pkg_upload"]] + 1
@@ -271,34 +274,36 @@ reweightViewServer <- function(id, user, decision_list, credentials) {
       all_pkgs <- dbSelect("SELECT DISTINCT name AS pkg_name FROM package")
       req(nrow(all_pkgs) > 0) # Stops re-weighting execution when no packages are in the database which crashes the app.
       
-      cmt_or_dec_pkgs <- unique(dplyr::bind_rows(
-        dbSelect("SELECT DISTINCT id AS pkg_name FROM comments where comment_type = 'o'"),
-        dbSelect("SELECT DISTINCT name AS pkg_name FROM package where decision_id IS NOT NULL")
-      ))
-      
-      cmt_or_dec_dropped_cmt <- " Since they may no longer be applicable, the final decision & comment have been dropped to allow for re-evaluation."
-      
-      # clear out any prior overall comments
-      dbUpdate("DELETE FROM comments WHERE comment_type = 'o'")
-      
-      for (i in 1:nrow(all_pkgs)) {
-        # insert comment for both mm and cum tabs
-        for (typ in c("mm","cum")) {
-          dbUpdate(
-            'INSERT INTO comments
-            VALUES({all_pkgs$pkg_name[i]}, {user$name}, {user$role},
+      if (input$reset_decisions) {
+        cmt_or_dec_pkgs <- unique(dplyr::bind_rows(
+          dbSelect("SELECT DISTINCT id AS pkg_name FROM comments where comment_type = 'o'"),
+          dbSelect("SELECT DISTINCT name AS pkg_name FROM package where decision_id IS NOT NULL")
+        ))
+        
+        cmt_or_dec_dropped_cmt <- " Since they may no longer be applicable, the final decision & comment have been dropped to allow for re-evaluation."
+        
+        # clear out any prior overall comments
+        dbUpdate("DELETE FROM comments WHERE comment_type = 'o'")
+        
+        for (i in 1:nrow(all_pkgs)) {
+          # insert comment for both mm and cum tabs
+          for (typ in c("mm","cum")) {
+            dbUpdate(
+              "INSERT INTO comments
+            VALUES({all_pkgs$pkg_name[i]}, {user$name}, {paste(user$role, collapse = ', ')},
             {paste0(weight_risk_comment(all_pkgs$pkg_name[i]), 
-                          ifelse(all_pkgs$pkg_name[i] %in% cmt_or_dec_pkgs$pkg_name, cmt_or_dec_dropped_cmt, ""))},
-            {typ}, {getTimeStamp()})'
-          )
+                          ifelse(all_pkgs$pkg_name[i] %in% cmt_or_dec_pkgs$pkg_name, cmt_or_dec_dropped_cmt, ''))},
+            {typ}, {getTimeStamp()})"
+            )
+          }
         }
-      }
-      
-      # Reset any decisions made prior to this.
-      pkg <- dbSelect("SELECT DISTINCT name AS pkg_name FROM package WHERE decision_id IS NOT NULL")
-      if (nrow(pkg) > 0) {
-        for (i in 1:nrow(pkg)) {
-          dbUpdate("UPDATE package SET decision_id = NULL where name = {pkg$pkg_name[i]}")
+        
+        # Reset any decisions made prior to this.
+        pkg <- dbSelect("SELECT DISTINCT name AS pkg_name FROM package WHERE decision_id IS NOT NULL")
+        if (nrow(pkg) > 0) {
+          for (i in 1:nrow(pkg)) {
+            dbUpdate("UPDATE package SET decision_id = NULL, decision_by = NULL, decision_date = NULL where name = {pkg$pkg_name[i]}")
+          }
         }
       }
